requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Pinecone (v5+)
pinecone-client==5.0.1

# Embeddings (choose one below)
sentence-transformers==3.0.1
torch>=2.2,<3.0

# PDF & DOCX parsing
pypdf==4.3.1
python-docx==1.1.2

# Groq SDK
groq==0.9.0

# Utils
pydantic-settings==2.4.0


langchain
langchain-community
openai
pinecone-client
python-docx
pypdf
voyageai
python-dotenv